+++ autologin_linux/test/autologin_linux_test.dart
@@ -87,6 +87,75 @@ void main() {
       },
     );
 
+    test(
+      'saveCredentials returns expected value',
+      () async {
+        autologin.setup(appId: 'demo.id', appName: 'Test app');
+        final report = await autologin.saveCredentials(SharedTests.expectedCredentials);
+        expect(report, equals(true));
+        expect(utils.log, [
+          isMethodCall(
+            'saveCredentials',
+            arguments: {
+              ...SharedTests.expectedCredentials.toJson(),
+              'appId': 'demo.id',
+              'appName': 'Test app',
+            },
+          ),
+        ]);
+      },
+    );
+
+    test(
+      'saveCredentials throws helpful errors',
+      () async {
+        const expectedMessage = "For the Linux platform you need to call AutologinPlugin.setup(appId: 'your.app.id', "
+            "appName: 'Your app name') before this call";
+
+        // All arguments missing
+        autologin.setup();
+        expect(
+          autologin.saveCredentials(SharedTests.expectedCredentials),
+          throwsA(
+            predicate(
+              (e) =>
+                  e is PlatformException &&
+                  e.message == expectedMessage &&
+                  e.details == "The missing arguments are 'appId' and 'appName'.",
+            ),
+          ),
+        );
+
+        // app name missing
+        autologin.setup(appId: 'demo.id');
+        expect(
+          autologin.saveCredentials(SharedTests.expectedCredentials),
+          throwsA(
+            predicate(
+              (e) =>
+                  e is PlatformException &&
+                  e.message == expectedMessage &&
+                  e.details == "The missing argument is 'appName'.",
+            ),
+          ),
+        );
+
+        // app id missing
+        autologin.setup(appName: 'test');
+        expect(
+          autologin.saveCredentials(SharedTests.expectedCredentials),
+          throwsA(
+            predicate(
+              (e) =>
+                  e is PlatformException &&
+                  e.message == expectedMessage &&
+                  e.details == "The missing argument is 'appId'.",
+            ),
+          ),
+        );
+      },
+    );
+
     test('saveLoginToken returns false', utils.saveLoginTokenReturnsFalse);
 
     test('requestLoginToken returns nothing', () async {
